pyproject.toml
[project]
name = "app"
version = "1.0.0"
description = "Users API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
